Cleaned up code. Properly used parser for more OOP
