+++ src/main/java/Duke/StringReplacer.java
@@ -1,0 +1,9 @@
+package duke;
+
+public class StringReplacer {
+
+    public String replaceAll(String input, String regex) {
+        return input.replaceAll(regex, "").trim();
+    }
+
+}
