+++ src/main/java/Duke/TaskList.java
@@ -13,7 +13,7 @@ public class TaskList {
     public List<Task> getList() {
         return list;
     }
-    
+
 
     /**
      * Prints tasks found when find is called
